requirements.txt
Django==1.9.5
django-authtools==1.5.0
django-braces==1.8.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-rest-swagger==0.3.5
djangorestframework==3.3.3
psycopg2==2.6.1
